backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
ultralytics
python-multipart
python-dotenv
google-genai
